requirements.txt
asttokens==2.4.1
backcall==0.2.0
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
certifi @ file:///croot/certifi_1700501669400/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cryptography @ file:///croot/cryptography_1702070282333/work
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
filelock @ file:///croot/filelock_1700591183607/work
fsspec==2023.12.2
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
idna @ file:///croot/idna_1666125576474/work
imageio==2.33.1
ipdb==0.13.13
ipython==8.20.0
jedi==0.19.1
Jinja2 @ file:///croot/jinja2_1666908132255/work
lazy_loader==0.3
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib-inline==0.1.6
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1690561992265/work
numpy @ file:///croot/numpy_and_numpy_base_1704311704800/work/dist/numpy-1.26.3-cp310-cp310-linux_x86_64.whl#sha256=a281f24b826e51f1c25bdd24960ab44b4bc294c65d81560441ba7fffd8ddd2a7
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow @ file:///croot/pillow_1696580024257/work
prompt-toolkit==3.0.43
protobuf==3.20.3
ptflops==0.7.2.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.17.2
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML @ file:///croot/pyyaml_1698096049011/work
requests @ file:///croot/requests_1690400202158/work
scikit-image==0.22.0
scipy==1.12.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
stack-data==0.6.3
sympy @ file:///croot/sympy_1701397643339/work
tensorboardX @ file:///tmp/build/80754af9/tensorboardx_1621440489103/work
tifffile==2023.12.9
tomli==2.0.1
torch==2.1.2
torchaudio==2.1.2
torchvideotransforms==0.1.2
torchvision==0.16.2
traitlets==5.14.1
triton==2.1.0
typing_extensions @ file:///croot/typing_extensions_1705599297034/work
tzdata==2023.4
urllib3 @ file:///croot/urllib3_1698257533958/work
wcwidth==0.2.13
